refactor: move version list component to page
