+++ src/components/index.js
@@ -3,4 +3,3 @@ export * from './Figure/Figure'
 export * from './FirstTimeSetup/FirstTimeSetup'
 export * from './Page/Page'
 export * from './UserGroupAccess/UserGroupAccess'
-export * from './VersionList/VersionList'
