+++ src/pages/ApkList/VersionList/AssignUserGroup.js
@@ -11,9 +11,9 @@ import {
 import isEmpty from 'lodash/isEmpty'
 import PropTypes from 'prop-types'
 import React, { useState } from 'react'
-import { useUpdateVersions, useUserGroups } from '../../hooks'
-import { prepareAPKListTable } from '../../pages/ApkList/helper'
-import { UserGroupAccess } from '../UserGroupAccess'
+import { UserGroupAccess } from '../../../components'
+import { useUpdateVersions, useUserGroups } from '../../../hooks'
+import { prepareAPKListTable } from '../helper'
 import { getGroup, updateList } from './helper'
 
 export const AssignUserGroup = ({ version, versionList, handleList }) => {
